+++ programs/store/src/ops/glv.rs
@@ -389,6 +389,7 @@ impl ExecuteGlvDepositOperation<'_, '_> {
                         deposit.tokens.initial_short_token.token(),
                     ),
                     None,
+                    true,
                 )?;
 
                 market_token_amount = market_token_amount